refactor(services): split GitHubService into query and auth modules
- Created electron/services/github/GitHubAuth.ts for authentication logic
- Created electron/services/github/GitHubQueries.ts for GraphQL queries
- Created electron/services/github/types.ts for shared type definitions
- Refactored GitHubService.ts to import from new modules
- Reduced GitHubService from 800 to 542 lines
- Removed unused prCheckCache variable
- Added token normalization (trim before storage)
- Added guard for empty candidates in buildBatchPRQuery
- Applied code formatting updates
