+++ electron/services/github/GitHubQueries.ts
@@ -1,0 +1,184 @@
+import type { PRCheckCandidate } from "./types.js";
+
+export const REPO_STATS_QUERY = `
+  query GetRepoStats($owner: String!, $repo: String!) {
+    repository(owner: $owner, name: $repo) {
+      issues(states: OPEN) { totalCount }
+      pullRequests(states: OPEN) { totalCount }
+    }
+  }
+`;
+
+export const LIST_ISSUES_QUERY = `
+  query GetIssues($owner: String!, $repo: String!, $states: [IssueState!], $cursor: String, $limit: Int = 20) {
+    repository(owner: $owner, name: $repo) {
+      issues(first: $limit, after: $cursor, states: $states, orderBy: {field: UPDATED_AT, direction: DESC}) {
+        pageInfo {
+          hasNextPage
+          endCursor
+        }
+        nodes {
+          number
+          title
+          url
+          state
+          updatedAt
+          author {
+            login
+            avatarUrl
+          }
+          assignees(first: 5) {
+            nodes {
+              login
+              avatarUrl
+            }
+          }
+          comments {
+            totalCount
+          }
+        }
+      }
+    }
+  }
+`;
+
+export const LIST_PRS_QUERY = `
+  query GetPRs($owner: String!, $repo: String!, $states: [PullRequestState!], $cursor: String, $limit: Int = 20) {
+    repository(owner: $owner, name: $repo) {
+      pullRequests(first: $limit, after: $cursor, states: $states, orderBy: {field: UPDATED_AT, direction: DESC}) {
+        pageInfo {
+          hasNextPage
+          endCursor
+        }
+        nodes {
+          number
+          title
+          url
+          state
+          isDraft
+          updatedAt
+          merged
+          author {
+            login
+            avatarUrl
+          }
+          reviews(first: 1) {
+            totalCount
+          }
+        }
+      }
+    }
+  }
+`;
+
+export const SEARCH_QUERY = `
+  query SearchItems($query: String!, $type: SearchType!, $cursor: String, $limit: Int = 20) {
+    search(query: $query, type: $type, first: $limit, after: $cursor) {
+      issueCount
+      pageInfo {
+        hasNextPage
+        endCursor
+      }
+      nodes {
+        ... on Issue {
+          number
+          title
+          url
+          state
+          updatedAt
+          author {
+            login
+            avatarUrl
+          }
+          assignees(first: 5) {
+            nodes {
+              login
+              avatarUrl
+            }
+          }
+          comments {
+            totalCount
+          }
+        }
+        ... on PullRequest {
+          number
+          title
+          url
+          state
+          isDraft
+          updatedAt
+          merged
+          author {
+            login
+            avatarUrl
+          }
+          reviews(first: 1) {
+            totalCount
+          }
+        }
+      }
+    }
+  }
+`;
+
+export function buildBatchPRQuery(
+  owner: string,
+  repo: string,
+  candidates: PRCheckCandidate[]
+): string {
+  const issueQueries: string[] = [];
+  const branchQueries: string[] = [];
+
+  for (let i = 0; i < candidates.length; i++) {
+    const candidate = candidates[i];
+
+    if (!candidate.issueNumber && !candidate.branchName) {
+      continue;
+    }
+
+    const alias = `wt_${i}`;
+
+    if (candidate.issueNumber) {
+      issueQueries.push(`
+        ${alias}_issue: repository(owner: "${owner}", name: "${repo}") {
+          issue(number: ${candidate.issueNumber}) {
+            timelineItems(itemTypes: [CROSS_REFERENCED_EVENT], last: 10) {
+              nodes {
+                ... on CrossReferencedEvent {
+                  source {
+                    ... on PullRequest {
+                      number
+                      url
+                      state
+                      isDraft
+                      merged
+                    }
+                  }
+                }
+              }
+            }
+          }
+        }
+      `);
+    }
+
+    if (candidate.branchName) {
+      const escapedBranch = JSON.stringify(candidate.branchName).slice(1, -1);
+      branchQueries.push(`
+        ${alias}_branch: repository(owner: "${owner}", name: "${repo}") {
+          pullRequests(first: 1, states: [OPEN, MERGED, CLOSED], headRefName: "${escapedBranch}", orderBy: {field: UPDATED_AT, direction: DESC}) {
+            nodes {
+              number
+              url
+              state
+              isDraft
+              merged
+            }
+          }
+        }
+      `);
+    }
+  }
+
+  return `query { ${issueQueries.join("\n")} ${branchQueries.join("\n")} }`;
+}
