+++ electron/services/github/index.ts
@@ -1,0 +1,20 @@
+export { GitHubAuth } from "./GitHubAuth.js";
+export type { GitHubTokenConfig, GitHubTokenValidation } from "./GitHubAuth.js";
+
+export {
+  REPO_STATS_QUERY,
+  LIST_ISSUES_QUERY,
+  LIST_PRS_QUERY,
+  SEARCH_QUERY,
+  buildBatchPRQuery,
+} from "./GitHubQueries.js";
+
+export type {
+  RepoContext,
+  RepoStats,
+  RepoStatsResult,
+  LinkedPR,
+  PRCheckResult,
+  PRCheckCandidate,
+  BatchPRCheckResult,
+} from "./types.js";
